Presidential Elections
I am suggesting that nominations be done by member states to increase the separation of powers. Nomination by petition would lead to hundreds of candidates running, while nomination by the Continental Congress would tie them too closely with their congressional allies. I think member states should nominate multiple candidates to increase diversity of choice.

The first round must happen to allow voters to know who the frontrunners are and to learn about them. The first two parts of the process produce candidates that are trusted outside of their member state, so that it cannot be skewed by the large states. The second two rounds ensure that there are genuine opposition candidates running in case no candidate is approved of by a majority. 

The second round is a Condorcet system. This means that if one candidate is preferred to every other candidate, they will win, even if very few people saw them as the best option. This is less certain in Instant Runoff Voting.