First Draft
I have written a first draft describing a Constitutional Court that requires a two-thirds majority to make rulings. This should prevent any one faction from controlling the court.

I have given most of the seats in the Constitutional Court to the State Governments. The rest are selected by a federal institution and ratified by the people. This should prevent the Court from tipping the balance of power too far towards the small states. I think we should require a two-thirds majority of the people for the justices to take office, to prevent these judges from becoming politicians.